Added optional maven-settings workspace

+++ scripts/maven-generate.sh
@@ -10,6 +10,15 @@ if [[ -f ${MAVEN_SETTINGS_FILE} ]]; then
     exit 0
 fi
 
+# Check if settings.xml exists in the workspace maven-settings
+if [[ -f "${WORKSPACES_MAVEN_SETTINGS_PATH}/settings.xml" ]]; then
+    cp "${WORKSPACES_MAVEN_SETTINGS_PATH}/settings.xml" "${MAVEN_SETTINGS_FILE}"
+    echo "Using '${MAVEN_SETTINGS_FILE}' copied from optional workspace maven-settings"
+    cat ${MAVEN_SETTINGS_FILE}
+    exit 0
+fi
+
+
 mkdir "${MAVEN_GENERATE_DIRECTORY}"
 
 cat > "${MAVEN_SETTINGS_FILE}" <<EOF
